backend/pest-service/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
python-dotenv
google-generativeai
